requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
beautifulsoup4==4.1.2
cssselect==0.9.1
fuzzywuzzy==0.2
htmlparser==0.0.2
nose==1.3.0
psycopg2==2.4.6
py==1.4.20
pygments==1.6
pymongo==2.6.3
pyodbc==3.0.6
pyopenssl==0.13.1
python-dateutil==2.1
requests==1.1.0
simplejson==3.0.7
sortedcollection==0.1.8.dev3
unidecode==0.04.14
version==0.1.0
w3lib==1.5
